pyproject.toml
[project]
name = "data-prep-kit-examples-dev-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arrow>=1.4.0",
    "data-prep-toolkit-transforms[language,ray]>=1.1.6",
    #"data-prep-toolkit-transforms[all,ray]>=1.1.6",
    "data-prep-toolkit[ray]>=1.1.6",
    "docling>=2.61.2",
    "litellm>=1.80.0",
    "llama-index>=0.11.23",
    "llama-index-embeddings-huggingface>=0.3.1",
    "llama-index-embeddings-litellm>=0.2.2",
    "llama-index-llms-litellm>=0.2.2",
    "llama-index-vector-stores-milvus>=0.2.7",
    "milvus-lite>=2.5.1",
    "pandas>=2.3.3",
    "pymilvus[model]>=2.6.3",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
